add CRUD for post on frontend

+++ backend/src/post/post.service.ts
@@ -1,4 +1,4 @@
-import {Injectable, NotFoundException} from '@nestjs/common';
+import {ForbiddenException, Injectable, NotFoundException} from '@nestjs/common';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import {InjectRepository} from "@nestjs/typeorm";
@@ -15,8 +15,15 @@ export class PostService {
       private repository: Repository<PostEntity>
   ) {}
 
-  create(dto: CreatePostDto) {
-    return this.repository.save(dto);
+  create(dto: CreatePostDto, userId: number) {
+    const firstParagraph = dto.body.find(obj => obj.type === 'paragraph')?.data?.text;
+    return this.repository.save({
+      title: dto.title,
+      body: dto.body,
+      tags: dto.tags,
+      user: { id: userId },
+      description: firstParagraph || '',
+    });
   }
 
   findAll() {
@@ -42,7 +49,7 @@ export class PostService {
   async search(dto: SearchPostDto) {
     const qb = this.repository.createQueryBuilder('p');
     qb.limit(dto.limit).offset(dto.offset);
-
+    qb.leftJoinAndSelect('p.user', 'user');
     if (dto.views) {
       qb.orderBy('views', dto.views);
     }
@@ -82,6 +89,7 @@ export class PostService {
     await this.repository
         .createQueryBuilder('posts')
         .whereInIds(id)
+        .leftJoinAndSelect('posts.user', 'user')
         .update()
         .set({
           views: () => 'views + 1'
@@ -91,21 +99,35 @@ export class PostService {
     return post;
   }
 
-  async update(id: number, dto: UpdatePostDto) {
+  async update(id: number, dto: UpdatePostDto, userId: number) {
     const post = await this.repository.findOne(+id);
 
     if (!post) {
       throw new NotFoundException('Статья не найдена');
     }
-
-    return this.repository.update(id, dto);
+    
+    if (post.user.id !== userId) {
+      throw new ForbiddenException('Нет доступа к этой статье');
+    }
+    
+    const firstParagraph = dto.body.find(obj => obj.type === 'paragraph')?.data?.text;
+    return this.repository.update(id, {
+      title: dto.title,
+      body: dto.body,
+      tags: dto.tags,
+      description: firstParagraph || '',
+    });
   }
 
-  async remove(id: number) {
+  async remove(id: number, userId: number) {
     const post = await this.repository.findOne(+id);
 
     if (!post) {
       throw new NotFoundException('Статья не найдена');
+    }
+  
+    if (post.user.id !== userId) {
+      throw new ForbiddenException('Нет доступа к этой статье');
     }
 
     return this.repository.delete(id);
